feat: simplification of domain entities and better separation of concerns
